Updating Historical Quote Day object to use a long data type for the Volume field as there are securities with volume that exceeds the data size of an int.

+++ TradierClient/Models/MarketData/HistoricalQuote.cs
@@ -25,6 +25,6 @@ namespace Tradier.Client.Models.MarketData
 
         [JsonProperty("close")] public float Close { get; set; }
 
-        [JsonProperty("volume")] public int Volume { get; set; }
+        [JsonProperty("volume")] public long Volume { get; set; }
     }
 }
